refactor: rename overview tab to evaluation across multiple components
