+++ src/front-end/typescript/lib/pages/proposal/team-with-us/view/tab/resource-questions.tsx
@@ -142,7 +142,7 @@ export function getResourceQuestionsOpportunityTab(
   return evaluating
     ? panelEvaluations.length
       ? ("consensus" as const)
-      : ("overview" as const)
+      : ("evaluation" as const)
     : ("resourceQuestions" as const);
 }
